Extract plane unit tests into plane_tests.rs
Tests were 29% of plane.rs (220 of 771 lines). Moving them to a sibling
file wired via `#[cfg(test)] #[path = "plane_tests.rs"] mod tests;` keeps
`use super::*;` reach to private items with zero visibility changes and
cuts plane.rs to 553 lines.

Also record the convention in CLAUDE.md so future unit-test modules land
in `<module>_tests.rs` by default.
