+++ CLAUDE.md
@@ -91,6 +91,7 @@ Workspace dependencies are defined in root `Cargo.toml` `[workspace.dependencies
 - Comments in English only
 - Authorization pattern: `admin.require_auth()` then `AccessControl::new(&e).assert_address_has_role(...)` or use helpers like `require_operations_admin_or_owner`
 - Events use `liquidity_pool_events`: `PoolEvents::new(&e).deposit_liquidity(...)`, `PoolEvents::new(&e).trade(...)`
+- **Keep unit tests in a sibling file, not inline.** When a source file (e.g. `foo.rs`) needs an in-crate unit-test module that accesses private items, put the tests in a sibling `foo_tests.rs` and wire them via `#[cfg(test)] #[path = "foo_tests.rs"] mod tests;` at the bottom of `foo.rs`. Do NOT write an inline `#[cfg(test)] mod tests { ... }` block. Reason: source files that bundle their tests inline grow large (tests often exceed 25% of file length), which inflates the context cost every time a code reader (human or agent) opens the file to study production logic. `#[path]` submodules keep `use super::*;` working against private items with zero visibility changes. Example: `plane.rs` + `plane_tests.rs`. Exception: a trivial test module (&lt;~40 lines, one or two tests) may stay inline.
 
 ## Development Workflow
